Completed Webpack, Bootstrap, Sass, and Font Awesome install with related dependencies.  Setup initial grid system and began populating restaurant details.
Next issue is to resolve a Font Awesome MIME issue that
does not appear to affect the page rendering.

Then I need to setup the menu tabs.

After that I need to polish the appearance and input
rendering logic so that items that a particular
restaurant does not have will not render.

Then I need to setup an API post to Alvin's module and
logic to handle the returned array from that POST
which will provide me with the raw materials to
calculate an individual restaurant's rating, the
number of reviews, and the most popular tags.