Solved node errors and other warnings from the prior commit comments by setting webpack to ignore server side components (lesson learned here is to formally separate those components from the beginning next time).
I am still getting an error from running
webpack and nodemon with a concurrently command
in npm start - there is no issue running them
separately so that will be my practice until I
have time to come back and address the issue.

The front end is also picking up the scss file
correctly now.  The only issue is that the custom
font is not loading correctly.  I can see it in
the browser but it is not appearing in the actual
text on the page - another issue to come back to -
the real issue is that bootstrap for some
inexplicable reasons does not permit the loading
of local fonts - unbelievable and reason enough
not to use it in my opinion.

Next task is to fill out the front end with the
remaining data and media.