Inputted additional nav bars at the top of the page
Inserted additional modules that appear below the
menus (at this point the only missing modules are
reservations, map, twitter, and save restaurant).

Also changed React rendering from jsx style to
using the createelement functionality in order to
use that on the proxy server.

Next step is to fix Container/Column widths in
bootstrap.

Then I need to refine the logic of the menu and
details modules to add conditional rendering